Add a global config file with configurable keybindings (#217)
Tab will look for a global config file at the following paths:
- `~/.config/tab.yml`
- `$CONFIG_DIR/tab.yml`, as provided by https://docs.rs/dirs/3.0.1/dirs/fn.config_dir.html

The global config file supports all properties that Workspace tab.ymls do, and it also supports a global 'key_bindings' property.  Here is an example:
```
key_bindings:
  - action: Disconnect
    keys: ctrl-A W

  - action: SelectInteractive
    keys: ctrl-T

workspace:
  - workspace: some-workspace/
  - tab: global-tab
```
The workspace definition will create a tab with the name of your home directory (e.g. `/Users/austinjones` creates `austinjones`).  The workspace tab and any defined workspace items will always be available when tab is invoked.

Workspace and repository definitions are resolved with your home directory as a base. E.g. `workspace: some-workspace` would resolve to `/Users/austinjones/some-workspace`

Key bindings are specified as a string of space-separated character codes.  Supported codes are:
- Visible ASCII characters (a-zA-Z0-9 and specials...)
- The range `ctrl-@` through `ctrl-_`, which includes ctrl-A to ctrl-Z
- The escape character as ESC

For single-character sequences, no stdin is sent to the connected shell.
For multi-character sequences, matched characters are sent to the shell, with the exception of the final character.